Updates
Created the profile page separated for each user, not only the existing editing profile page. Made the follow/unfollow button to work, now the users can follow each other on the platform. Updated the post pages now you can see likes and the category it is a child of. Now if the user is not logged in they have the option to create or login in an account hiding the create post button if the user is not logged in. The crate account how has the option to choose your username as well.