Filters
Now you can filter the posts by categories and access the categories as well from within the posts. The posts are available to be read by non authenticated users as well. Made a few tweaks to the css so that some buttons and texts are looking better on light and dark theme.